app/bot/filters/admin_filter.py
from aiogram.filters import BaseFilter
from aiogram.types import Message, CallbackQuery

from dotenv import load_dotenv, find_dotenv
import os

load_dotenv(find_dotenv())

ADMIN_ID = int(os.getenv("ADMIN_ID"))

class IsAdmin(BaseFilter):
    async def __call__(self, event: Message | CallbackQuery) -> bool:
        return event.from_user.id == ADMIN_ID
    
